+++ changelog.md
@@ -1,6 +1,11 @@
 # Changelog
 
-- * 2620093 - 2018-10-27: Release v9.4.1 
+- * 9d1c1e7 - 2018-10-28: Release v9.4.2 
+- * 1ad42fe - 2018-10-28: Clears name cache with indexerid and show name 
+- * e26ad3e - 2018-10-28: Converted Timezone Updater to class 
+- * 7e3baa8 - 2018-10-28: Disabled warning for when no nzb/torrent provider is picked if no nzb/torrent client enabled. Disabled search column for display show view if no nzb/torrent is picked. 
+- * 9969a0b - 2018-10-28: Added function to strip accents to encoding functions 
+- * bf93454 - 2018-10-27: Release v9.4.1 
 - * 3dbd257 - 2018-10-27: Release v9.3.100 
 - * 479d80e - 2018-10-27: Removed app_id and replaced with app_sub. App now links its self to SSO user via userid instead of appid. SSL cert and key now stored/retrieved from data folder. 
 - * 33ea2f1 - 2018-10-21: Release v9.3.99